Added few new controllers to distribute load of admin controller.
